Change the Cart's output filename from "OGPDownload.zip" to "FGDL_Download_<YYYYMONDD>.zip"
On the map - the Find place box is hard to see. Make the background less
transparent.

On the map -- add "Zoom to Florida extent" button next to "Zoom to
global extent"

In download select format dropdown, change "ArcGIS Layer File" to "Map
Service for ArcGIS"

When preview is clicked, auto-expand the record to see the legend and
identify

Change "i" icon column heading (for metadata) to "Meta" 

Change "Name" column heading to "Title"

Switch positions of Preview checkbox and Add to Cart checkboxes